python/ambar/requirements.txt
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
Jinja2==2.10
MarkupSafe==1.1.1
PyYAML==5.1
Werkzeug==0.15.1
aniso8601==6.0.0
astroid==1.6.5
certifi==2019.3.9
chardet==3.0.4
idna==2.8
isort==4.3.16
itsdangerous==1.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==1.9.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
wrapt==1.11.1
